Updated activity log table stub to fix multiple runs of 'migrate:fresh' issue

+++ database/migrations/create_activity_log_table.php.stub
@@ -8,20 +8,24 @@ class CreateActivityLogTable extends Migration
 {
     public function up()
     {
-        Schema::connection(config('activitylog.database_connection'))->create(config('activitylog.table_name'), function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->string('log_name')->nullable();
-            $table->text('description');
-            $table->nullableMorphs('subject', 'subject');
-            $table->nullableMorphs('causer', 'causer');
-            $table->json('properties')->nullable();
-            $table->timestamps();
-            $table->index('log_name');
-        });
+        if (!Schema::connection(config('activitylog.database_connection'))->hasTable(config('activitylog.table_name'))) {
+            Schema::connection(config('activitylog.database_connection'))->create(config('activitylog.table_name'), function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->string('log_name')->nullable();
+                $table->text('description');
+                $table->nullableMorphs('subject', 'subject');
+                $table->nullableMorphs('causer', 'causer');
+                $table->json('properties')->nullable();
+                $table->timestamps();
+                $table->index('log_name');
+            });
+        }
     }
 
     public function down()
     {
-        Schema::connection(config('activitylog.database_connection'))->dropIfExists(config('activitylog.table_name'));
+        if (Schema::connection(config('activitylog.database_connection'))->hasTable(config('activitylog.table_name'))) {
+            Schema::connection(config('activitylog.database_connection'))->dropIfExists(config('activitylog.table_name'));
+        }
     }
 }
